requirements.txt
aiogram>=3.4,<4
python-dotenv>=1.0
pytest>=8.0
